Add files via upload
Task 2 
Write a Python script that uses a function or functions to implement an elimination game. The 
game starts with a team of 12 players. During the play, randomly eliminate between 2 – 6 
players. Each time a player is eliminated, output the name of the player with a send-forth 
message. Wait for 30 seconds before the next round.  
 
Note your core function should accept two parameters. The first parameter should be a list data 
type for holding the players and the second parameter should be an integer value denoting the 
index of a player to eliminate. Use a random generator to generate the indexes within a range 
of 1 to 6. In each case, remove a player from the given list (first parameter) based on the 
randomly generated index. The function should return the resulting list as a tuple. 
 
Interactively create a list of 12 player names. Then request a user to enter the number of players 
to be eliminated. Remember that at max 6 players could be eliminated. Ensure that this range 
is enforced and handle exceptions. Invoke your function with the appropriate parameters. At 
the end of the program, output per line, first the returned tuple of remaining players and then 
the original set of players.